Merge(운동기록) : 테스트코 병합
